# docs/swagger.yaml
basePath: /
definitions:
...
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.Anomalia:
    properties:
      accion_sugerida:
        type: string
      anomalia_id:
        type: integer
      anomalia_referencia_id:
        type: integer
      camion_id:
        type: integer
      categoria_clasificada:
        type: string
      ciudadano_id:
        description: |-
          CiudadanoID: quien reporto, cuando fue un ciudadano (nunca junto con
          ConductorID -- ver CreateAnomaliaController.go, que deriva cual de los
          dos llenar segun el role_id del JWT, no segun lo que mande el cliente).
        type: integer
      conductor_id:
        type: integer
      created_at:
        type: string
      descripcion:
        type: string
      eliminado:
        type: boolean
      estado:
        type: string
      estado_pipeline:
        description: |-
          Pipeline modelo_reportes -> clasificador_reportes. Se llenan por
          ActualizarPipeline en background, no al crear la anomalia (por eso
          EstadoPipeline arranca en "pendiente" via DEFAULT de la columna).
        type: string
      fecha_reporte:
        type: string
      fecha_resolucion:
        type: string
      inferencia_id:
        type: integer
      json_ruta:
        type: string
      lat:
        description: |-
          Lat/Lon: coordenadas GPS de donde ocurrio la anomalia (no necesariamente
          coinciden con un punto_recoleccion existente -- un "calle bloqueada"
          puede pasar en cualquier punto de una ruta). Nullable: no todo cliente
          puede capturar ubicacion todavia. Mismo nombre/tipo que
          Rutas/domain/entities/PuntoRecoleccion.go (Lat/Lon float64) para
          mantener la convencion, pero como puntero porque aqui es opcional.
          Insumo para el algoritmo genetico de rutas (AG): necesita saber donde
          esta el bloqueo/incidente para decidir que arista del grafo modificar.
        type: number
      lon:
        type: number
      nivel_riesgo:
        type: string
      pipeline_error:
        type: string
      pipeline_intentos:
        description: |-
          PipelineIntentos cuenta cuantas veces se corrio/reclamo el pipeline
          sobre esta fila. Lo usa PipelineRetryWorker para acotar los reintentos
          automaticos de un estado_pipeline = 'error'.
        type: integer
      punto_id:
        type: integer
      ruta_id:
        type: integer
      subtipo_clasificado:
        type: string
      tipo_anomalia:
        $ref: '#/definitions/entities.TipoAnomalia'
      updated_at:
        type: string
    type: object
  entities.AnomaliaListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.Anomalia'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.AnomaliaMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.AnomaliaResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Anomalia'
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.Camion:
    properties:
      camion_id:
        type: integer
      color_disponibilidad:
        type: string
      created_at:
        type: string
      disponibilidad_id:
        type: integer
      es_rentado:
        type: boolean
      modelo:
        type: string
      nombre_disponibilidad:
        type: string
      placa:
        type: string
      tipo_camion_id:
        type: integer
      updated_at:
        type: string
    type: object
  entities.CamionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Camion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.CreateAnomaliaRequest:
    properties:
      anomalia_referencia_id:
        type: integer
      camion_id:
        type: integer
      ciudadano_id:
        type: integer
      conductor_id:
        description: |-
          ConductorID/CiudadanoID: solo staff puede fijarlos libremente aqui. Si
          quien crea es ciudadano o conductor, el backend ignora estos dos
          campos y deriva el que corresponda del JWT (ver CreateAnomaliaController.go).
        type: integer
      descripcion:
        type: string
      estado:
        type: string
      fecha_reporte:
        description: 'Formato ISO 8601: YYYY-MM-DDTHH:MM:SSZ'
        example: "2026-07-22T19:30:00Z"
        type: string
      json_ruta:
        type: string
      lat:
        description: |-
          Coordenadas GPS de donde ocurrio la anomalia. Opcionales, insumo futuro
          del algoritmo genetico de rutas.
        type: number
      lon:
        type: number
      punto_id:
        type: integer
      ruta_id:
        type: integer
      tipo_anomalia:
        description: 'Valores validos: ANOMALIA | INCIDENCIA | REPORTE_CONDUCTOR |
          REPORTE_FALLA_CRITICA | SEGUIMIENTO_FALLA_CRITICA'
        example: REPORTE_CONDUCTOR
        type: string
    required:
    - descripcion
    - fecha_reporte
    - tipo_anomalia
    type: object
  entities.CreateCamionRequest:
    properties:
      disponibilidad_id:
        type: integer
      es_rentado:
        type: boolean
      modelo:
        type: string
      placa:
        type: string
      tipo_camion_id:
        type: integer
    required:
    - disponibilidad_id
    - modelo
    - placa
    - tipo_camion_id
    type: object
...
  entities.CreateRutaRequest:
    properties:
      descripcion:
        type: string
      json_ruta: {}
      nombre:
        type: string
    required:
    - json_ruta
    - nombre
    type: object
...
  entities.EstadoCamion:
    properties:
      camion_id:
        type: integer
      estado:
        description: nullable
        type: string
      estado_id:
        type: integer
      observaciones:
        description: nullable
        type: string
      timestamp:
        description: nullable
        type: string
    type: object
  entities.EstadoCamionListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.EstadoCamion'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.Ruta:
    properties:
      conductor_id:
        description: |-
          ConductorID es el id_chofer con asignación activa (historial_asignacion_camion)
          del camión actualmente asignado a la ruta (ruta_camion). Solo lo llena GetActivas.
          swagger:description ID del conductor asignado (null si no hay asignación activa)
        example: 12
        type: integer
      created_at:
        type: string
      descripcion:
        example: Recolección zona centro
        type: string
      eliminado:
        example: false
        type: boolean
      json_ruta:
        type: string
      nombre:
        example: Ruta Centro
        type: string
      ruta_id:
        example: 1
        type: integer
    type: object
...
  entities.RutaResponse:
    properties:
      data:
        $ref: '#/definitions/entities.Ruta'
      success:
        type: boolean
    type: object
...
  entities.TipoAnomalia:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - TipoAnomaliaAnomalia
    - TipoAnomaliaIncidencia
    - TipoAnomaliaReporteConductor
    - TipoAnomaliaReporteFallaCritica
    - TipoAnomaliaSeguimientoFallaCritica
  entities.UpdateAnomaliaRequest:
    properties:
      anomalia_referencia_id:
        type: integer
      camion_id:
        type: integer
      ciudadano_id:
        type: integer
      conductor_id:
        type: integer
      descripcion:
        type: string
      eliminado:
        type: boolean
      estado:
        type: string
      fecha_reporte:
        example: "2026-07-22T19:30:00Z"
        type: string
      fecha_resolucion:
        type: string
      json_ruta:
        type: string
      lat:
        type: number
      lon:
        type: number
      punto_id:
        type: integer
      ruta_id:
        type: integer
      tipo_anomalia:
        example: REPORTE_CONDUCTOR
        type: string
    required:
    - descripcion
    - fecha_reporte
    - tipo_anomalia
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/anomalias/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar anomalías
      tags:
      - Anomalia
    post:
      description: |-
        Abierto a cualquier usuario autenticado (ciudadano, conductor
        o staff) -- solo el resto del CRUD de anomalias queda
        restringido a ADMIN/SUPERVISOR/COORDINADOR.
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateAnomaliaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear anomalía
      tags:
      - Anomalia
  /api/anomalias/{id}:
    delete:
      description: |-
        Staff puede eliminar cualquier anomalia. Un conductor o
        ciudadano solo puede eliminar un reporte propio
        (conductor_id/ciudadano_id == su user_id, segun corresponda).
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar anomalía
      tags:
      - Anomalia
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Anomalía por ID
      tags:
      - Anomalia
    put:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateAnomaliaRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar anomalía
      tags:
      - Anomalia
...
  /api/anomalias/referencia/{referenciaId}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Anomalías por anomalía de referencia (seguimientos)
      tags:
      - Anomalia
...
  /api/camion/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar camiones
      tags:
      - Camion
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateCamionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.CamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear camión
      tags:
      - Camion
...
  /api/rutas/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar rutas
      tags:
      - Ruta
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateRutaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.RutaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear ruta
      tags:
      - Ruta
...
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"